Fix an if that didn't include all its statement
Previously the compiler didn't complain about it because the second
statement used tabs instead of spaces, so it thought that the indentation
was different. After replacing tabs by spaces, the compiler knows that
this code is misleading.

It looks like the intent of the code was to include both statements in the
conditional. This effect only needs to be executed in the first tick of
the row.